Build out chat interface to replace AgentThinking and respond to every command
