--- droidlet/dashboard/web/src/components/Interact/Message.css
@@ -1,156 +1,0 @@
-/*
-Copyright (c) Facebook, Inc. and its affiliates.
-*/
-@import url("https://fonts.googleapis.com/css?family=Red+Hat+Display:400,500,900&display=swap");
-body,
-html {
-  font-family: Red hat Display, sans-serif;
-  font-weight: 400;
-  line-height: 1.25em;
-  letter-spacing: 0.025em;
-  color: #333;
-  background: #f7f7f7;
-}
-
-.center {
-  position: absolute;
-  top: 55%;
-  left: 50%;
-  height: 85%;
-  width: 95%;
-  transform: translate(-50%, -50%);
-}
-
-.chat {
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: space-between;
-  width: 100%;
-  height: 100%;
-  z-index: 2;
-  box-sizing: border-box;
-  border-radius: 1rem;
-  background: white;
-  box-shadow: 0 0 8rem 0 rgba(0, 0, 0, 0.1),
-    0rem 2rem 4rem -3rem rgba(0, 0, 0, 0.5);
-}
-.chat .time {
-  font-size: 0.8rem;
-  background: #eee;
-  padding: 0.25rem 1rem;
-  border-radius: 2rem;
-  color: #999;
-  width: -webkit-fit-content;
-  width: -moz-fit-content;
-  width: fit-content;
-  margin: 0 auto;
-  z-index: 3;
-}
-.chat .messages {
-  position: absolute;
-  bottom: 4.5rem;
-  margin-top: 1rem;
-  width: 100%;
-  height: calc(100% - 4.5rem);
-  background: #f7f7f7;
-  overflow-y: auto;
-  border-top-right-radius: 1rem;
-  border-top-left-radius: 1rem;
-  box-shadow: inset 0 2rem 2rem -2rem rgba(0, 0, 0, 0.05),
-    inset 0 -2rem 2rem -2rem rgba(0, 0, 0, 0.05);
-}
-.chat .messages .messagelist {
-  position: absolute;
-  bottom: 0;
-  width: 100%;
-  list-style-type: none;
-  padding-left: 0;
-  margin: 0;
-}
-.chat .messages .message {
-  box-sizing: border-box;
-  padding: 0.5rem 1rem;
-  margin: 1rem;
-  background: #fff;
-  border-radius: 1.125rem 1.125rem 1.125rem 0;
-  min-height: 2.25rem;
-  width: -webkit-fit-content;
-  width: -moz-fit-content;
-  width: fit-content;
-  max-width: 66%;
-  box-shadow: 0 0 2rem rgba(0, 0, 0, 0.075),
-    0rem 1rem 1rem -1rem rgba(0, 0, 0, 0.1);
-  color: black;
-  text-align: left;
-}
-.chat .messages .message.user {
-  margin: 1rem 1rem 1rem auto;
-  border-radius: 1.125rem 1.125rem 0 1.125rem;
-  background: #333;
-  color: white;
-  text-align: right;
-}
-.chat .input {
-  position: absolute;
-  bottom: 0;
-  box-sizing: content-box;
-  min-width: 85%;
-  margin: auto;
-  flex-basis: auto;
-  display: flex;
-  align-items: center;
-  padding: 1.25rem;
-}
-.chat .input i {
-  font-size: 1.5rem;
-  color: #666;
-  cursor: pointer;
-  transition: color 200ms;
-}
-.chat .input i:hover {
-  color: #333;
-}
-.chat .input input {
-  border: none;
-  background-image: none;
-  background-color: white;
-  padding: 0.5rem 1rem;
-  border-radius: 1.125rem;
-  flex-grow: 2;
-  box-shadow: 0 0 1rem rgba(0, 0, 0, 0.1),
-    0rem 1rem 1rem -1rem rgba(0, 0, 0, 0.2);
-  font-family: Red hat Display, sans-serif;
-  font-weight: 400;
-  letter-spacing: 0.025em;
-}
-.chat .input input:placeholder {
-  color: #999;
-}
-
-@-webkit-keyframes typing {
-  0%,
-  75%,
-  100% {
-    transform: translate(0, 0.25rem) scale(0.9);
-    opacity: 0.5;
-  }
-  25% {
-    transform: translate(0, -0.25rem) scale(1);
-    opacity: 1;
-  }
-}
-
-@keyframes typing {
-  0%,
-  75%,
-  100% {
-    transform: translate(0, 0.25rem) scale(0.9);
-    opacity: 0.5;
-  }
-  25% {
-    transform: translate(0, -0.25rem) scale(1);
-    opacity: 1;
-  }
-}
